requirement1.txt
asgiref==3.4.1
Django==3.0.2
django-crispy-forms==1.13.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.0.0
Pillow==8.4.0
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
